Public API updates
