+++ src/Http/Http.Abstractions/src/HttpResults/INestedHttpResult.cs
@@ -1,0 +1,16 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.AspNetCore.Http;
+
+/// <summary>
+/// Defines a contract that represents the result of an HTTP endpoint
+/// that contains a nested <see cref="IResult"/> type.
+/// </summary>
+public interface INestedHttpResult
+{
+    /// <summary>
+    /// Gets the actual <see cref="IResult"/> returned by the <see cref="Endpoint"/> route handler delegate.
+    /// </summary>
+    IResult Result { get; }
+}
