+++ src/Http/Http.Abstractions/src/HttpResults/IResult.cs
