+++ src/Http/Http.Abstractions/src/PublicAPI.Unshipped.txt
@@ -11,7 +11,20 @@ Microsoft.AspNetCore.Http.EndpointMetadataCollection.Enumerator.Current.get -> o
 Microsoft.AspNetCore.Http.EndpointMetadataCollection.GetRequiredMetadata<T>() -> T!
 Microsoft.AspNetCore.Http.IBindableFromHttpContext<TSelf>
 Microsoft.AspNetCore.Http.IBindableFromHttpContext<TSelf>.BindAsync(Microsoft.AspNetCore.Http.HttpContext! context, System.Reflection.ParameterInfo! parameter) -> System.Threading.Tasks.ValueTask<TSelf?>
+Microsoft.AspNetCore.Http.IContentTypeHttpResult
+Microsoft.AspNetCore.Http.IContentTypeHttpResult.ContentType.get -> string?
+Microsoft.AspNetCore.Http.IFileHttpResult
+Microsoft.AspNetCore.Http.IFileHttpResult.ContentType.get -> string?
+Microsoft.AspNetCore.Http.IFileHttpResult.FileDownloadName.get -> string?
+Microsoft.AspNetCore.Http.INestedHttpResult
+Microsoft.AspNetCore.Http.INestedHttpResult.Result.get -> Microsoft.AspNetCore.Http.IResult!
 Microsoft.AspNetCore.Http.IRouteHandlerFilter.InvokeAsync(Microsoft.AspNetCore.Http.RouteHandlerInvocationContext! context, Microsoft.AspNetCore.Http.RouteHandlerFilterDelegate! next) -> System.Threading.Tasks.ValueTask<object?>
+Microsoft.AspNetCore.Http.IStatusCodeHttpResult
+Microsoft.AspNetCore.Http.IStatusCodeHttpResult.StatusCode.get -> int
+Microsoft.AspNetCore.Http.IValueHttpResult
+Microsoft.AspNetCore.Http.IValueHttpResult.RawValue.get -> object?
+Microsoft.AspNetCore.Http.IValueHttpResult<TValue>
+Microsoft.AspNetCore.Http.IValueHttpResult<TValue>.Value.get -> TValue?
 Microsoft.AspNetCore.Http.Metadata.IFromFormMetadata
 Microsoft.AspNetCore.Http.Metadata.IFromFormMetadata.Name.get -> string?
 Microsoft.AspNetCore.Http.Metadata.IRequestSizeLimitMetadata
