+++ src/Http/Http.Results/src/PushStreamHttpResult.cs
@@ -11,7 +11,7 @@ namespace Microsoft.AspNetCore.Http.HttpResults;
 /// Represents an <see cref="IResult"/> that when executed will
 /// write a file from the writer callback to the response.
 /// </summary>
-public sealed class PushStreamHttpResult : IResult, IFileHttpResult
+public sealed class PushStreamHttpResult : IResult, IFileHttpResult, IContentTypeHttpResult
 {
     private readonly Func<Stream, Task> _streamWriterCallback;
